Comment system complete, Quiz Taken mostly complete

+++ server/src/graphql/commentSchema.ts
@@ -1,5 +1,14 @@
 import { GraphQLObjectType, GraphQLSchema, GraphQLID, GraphQLString, GraphQLInt, GraphQLBoolean, GraphQLList, GraphQLNonNull } from 'graphql';
-import { Comment } from '../database/comment.js';
+import { Comment, User } from '../database/index';
+
+const Username = new GraphQLObjectType({
+    name: 'Username',
+    fields: () => ({
+        id: {type: GraphQLString},
+        username: {type: GraphQLString},
+        editable: {type: GraphQLBoolean}
+    })
+})
 
 export const CommentType = new GraphQLObjectType({
     name: 'Comment',
@@ -7,7 +16,20 @@ export const CommentType = new GraphQLObjectType({
         id: { type: GraphQLString },
         createdAt: { type: new GraphQLNonNull(GraphQLString) },
         contents: { type: new GraphQLNonNull(GraphQLString) },
-        infoNodeId: { type: GraphQLString }
+        infoNodeId: { type: GraphQLString },
+        userID: { 
+            type: Username,
+            async resolve(parent: any, args: any, context: any) {
+                const token = context.req.signedCookies["jwt"];
+                const auth = AuthenticationController.authenticateJWT(token);
+                const temp = await User.findById(parent.userID);
+                if(auth.valid) {
+                    return {id: temp._id, username: temp.username, editable: auth.userID == temp._id};
+                } else {
+                    return {id: temp._id, username: temp.username, editable: false};
+                }
+            }
+        }
     })
 });
 
@@ -27,35 +49,58 @@ export const CommentQueries = {
     }
 };
 
+import {AuthenticationController} from './../controllers/authentication';
 export const CommentMutations = {
     addComment: {
         type: CommentType,
         args: {
             contents: { type: new GraphQLNonNull(GraphQLString) },
-            infoNodeId: { type: new GraphQLNonNull(GraphQLString) }
+            infoNodeId: { type: new GraphQLNonNull(GraphQLString) },
         },
-        resolve(parent: any, args: any) {
-            const comment = new Comment(args);
-            return comment.save();
+        resolve(parent: any, args: any, context: any) {
+            const token = context.req.signedCookies["jwt"];
+            const auth = AuthenticationController.authenticateJWT(token);
+            if(auth.valid) {
+                const comment = new Comment({...args, userID: auth.userID});
+                return comment.save();
+            }
+            throw new Error();
         }
     },
     updateComment: {
         type: CommentType,
         args: {
-            id: { type: GraphQLString },
-            createdAt: { type: new GraphQLNonNull(GraphQLString) },
-            contents: { type: new GraphQLNonNull(GraphQLString) },
-            infoNodeId: { type: new GraphQLNonNull(GraphQLString) }
+            id: {type: new GraphQLNonNull(GraphQLString)},
+            contents: {type: new GraphQLNonNull(GraphQLString)},
+            userID: {type: new GraphQLNonNull(GraphQLString)}
         },
-        resolve(parent: any, args: any) {
-            return Comment.findByIdAndUpdate(args.id, args);
+        async resolve(parent: any, args: any, context: any) {
+            const token = context.req.signedCookies["jwt"];
+            const auth = AuthenticationController.authenticateJWT(token);
+            if(auth.valid) {
+                if(auth.userID == args.userID) {
+                    return Comment.findByIdAndUpdate(args.id, {contents: args.contents}, {new: true});
+                }
+            }
+            throw new Error();
         }
     },
     deleteComment: {
         type: CommentType,
-        args: { id: { type: GraphQLString}},
-        resolve(parent: any, args: any) {
-            return Comment.findByIdAndRemove(args.id);
-        }
+        args: { 
+            id: { type: GraphQLString},
+            userID: {type: new GraphQLNonNull(GraphQLString)} 
+        },
+        resolve(parent: any, args: any, context: any) {
+            const token = context.req.signedCookies["jwt"];
+            const auth = AuthenticationController.authenticateJWT(token);
+            if(auth.valid) {
+                if(auth.userID == args.userID) {
+                    return Comment.findByIdAndRemove(args.id);
+                }
+            }
+            throw new Error();
+        },
+        
     }
 }
